move sourcetype identification logic to sourcetype package
