+++ internal/pkg/source/source.go
@@ -16,13 +16,13 @@
 package source
 
 import (
-	"fmt"
 	"go/token"
 	"go/types"
 
 	"github.com/google/go-flow-levee/internal/pkg/config"
 	"github.com/google/go-flow-levee/internal/pkg/fieldpropagator"
 	"github.com/google/go-flow-levee/internal/pkg/fieldtags"
+	"github.com/google/go-flow-levee/internal/pkg/sourcetype"
 	"github.com/google/go-flow-levee/internal/pkg/utils"
 	"golang.org/x/tools/go/analysis/passes/buildssa"
 	"golang.org/x/tools/go/ssa"
@@ -90,7 +90,7 @@ func identify(conf *config.Config, ssaInput *buildssa.SSA, taggedFields fieldtag
 func sourcesFromParams(fn *ssa.Function, conf *config.Config, taggedFields fieldtags.ResultType) []*Source {
 	var sources []*Source
 	for _, p := range fn.Params {
-		if IsSourceType(conf, taggedFields, p.Type()) {
+		if sourcetype.IsSourceType(conf, taggedFields, p.Type()) {
 			sources = append(sources, New(p))
 		}
 	}
@@ -106,7 +106,7 @@ func sourcesFromClosures(fn *ssa.Function, conf *config.Config, taggedFields fie
 	for _, p := range fn.FreeVars {
 		switch t := p.Type().(type) {
 		case *types.Pointer:
-			if IsSourceType(conf, taggedFields, t) {
+			if sourcetype.IsSourceType(conf, taggedFields, t) {
 				sources = append(sources, New(p))
 			}
 		}
@@ -135,19 +135,19 @@ func isSourceNode(n ssa.Node, conf *config.Config, propagators fieldpropagator.R
 
 	// Values produced by sanitizers are not sources.
 	case *ssa.Alloc:
-		return !isProducedBySanitizer(v, conf) && IsSourceType(conf, taggedFields, n.(ssa.Value).Type())
+		return !isProducedBySanitizer(v, conf) && sourcetype.IsSourceType(conf, taggedFields, n.(ssa.Value).Type())
 
 	// Values produced by sanitizers are not sources.
 	// Values produced by field propagators are.
 	case *ssa.Call:
 		return !isProducedBySanitizer(v, conf) &&
-			(propagators.IsFieldPropagator(v) || IsSourceType(conf, taggedFields, n.(ssa.Value).Type()))
+			(propagators.IsFieldPropagator(v) || sourcetype.IsSourceType(conf, taggedFields, n.(ssa.Value).Type()))
 
 	// A type assertion can assert that an interface is of a source type.
 	// Only panicky type asserts will refer to the source Value.
 	// The typed value returned in (value, ok) type assertions are examined in the case for ssa.Extract instructions.
 	case *ssa.TypeAssert:
-		return !v.CommaOk && IsSourceType(conf, taggedFields, v.AssertedType)
+		return !v.CommaOk && sourcetype.IsSourceType(conf, taggedFields, v.AssertedType)
 
 	// An Extract is used to obtain a value from an instruction that returns multiple values.
 	// If the extracted value is a Pointer to a Source, it won't have an Alloc, so we need to
@@ -155,11 +155,11 @@ func isSourceNode(n ssa.Node, conf *config.Config, propagators fieldpropagator.R
 	case *ssa.Extract:
 		t := v.Tuple.Type().(*types.Tuple).At(v.Index).Type()
 		_, ok := t.(*types.Pointer)
-		return ok && IsSourceType(conf, taggedFields, t)
+		return ok && sourcetype.IsSourceType(conf, taggedFields, t)
 
 	// Unary operator <- can receive sources from a channel.
 	case *ssa.UnOp:
-		return v.Op == token.ARROW && IsSourceType(conf, taggedFields, n.(ssa.Value).Type())
+		return v.Op == token.ARROW && sourcetype.IsSourceType(conf, taggedFields, n.(ssa.Value).Type())
 
 	// Field access (Field, FieldAddr),
 	// collection access (Index, IndexAddr, Lookup),
@@ -168,53 +168,8 @@ func isSourceNode(n ssa.Node, conf *config.Config, propagators fieldpropagator.R
 	case *ssa.Field, *ssa.FieldAddr,
 		*ssa.Index, *ssa.IndexAddr, *ssa.Lookup,
 		*ssa.MakeMap, *ssa.MakeChan:
-		return IsSourceType(conf, taggedFields, n.(ssa.Value).Type())
+		return sourcetype.IsSourceType(conf, taggedFields, n.(ssa.Value).Type())
 	}
-}
-
-// IsSourceType determines whether a Type is a Source Type.
-// A Source Type is either:
-// - A Named Type that is classified as a Source
-// - A composite type that contains a Source Type
-// - A Struct Type that contains a tagged field
-func IsSourceType(c *config.Config, tf fieldtags.ResultType, t types.Type) bool {
-	deref := utils.Dereference(t)
-	switch tt := deref.(type) {
-	case *types.Named:
-		return c.IsSourceType(utils.DecomposeType(tt)) || IsSourceType(c, tf, tt.Underlying())
-	case *types.Array:
-		return IsSourceType(c, tf, tt.Elem())
-	case *types.Slice:
-		return IsSourceType(c, tf, tt.Elem())
-	case *types.Chan:
-		return IsSourceType(c, tf, tt.Elem())
-	case *types.Map:
-		key := IsSourceType(c, tf, tt.Key())
-		elem := IsSourceType(c, tf, tt.Elem())
-		return key || elem
-	case *types.Struct:
-		return hasTaggedField(tf, tt)
-	case *types.Basic, *types.Tuple, *types.Interface, *types.Signature:
-		// These types do not currently represent possible source types
-		return false
-	case *types.Pointer:
-		// This should be unreachable due to the dereference above
-		return false
-	default:
-		// The above should be exhaustive.  Reaching this default case is an error.
-		fmt.Printf("unexpected type received: %T %v; please report this issue\n", tt, tt)
-		return false
-	}
-}
-
-func hasTaggedField(taggedFields fieldtags.ResultType, s *types.Struct) bool {
-	for i := 0; i < s.NumFields(); i++ {
-		f := s.Field(i)
-		if taggedFields.IsSource(f) {
-			return true
-		}
-	}
-	return false
 }
 
 func isProducedBySanitizer(v ssa.Value, conf *config.Config) bool {
